Second Project
I used the alt key to put down multiple cursors so I was able to type in all the list items and hyperlinks and heading all at the same time after I had copied the text into it. I ask someone next to me about the ID attribute then I looked it up as well then Max stood up and told us all again how to properly set up that attribute and then I did it correctly and check to make sure it worked.  I added in the mail but the subject part alluded me for a while until i asked Ryan about it and he helped me. Ryan also helped me with the link to a certain point in the different site.